GH-99 - moved mixtape menu to separate component in order to reuse functionality
