+++ frontend/src/components/MixtapeCard.tsx
@@ -2,25 +2,12 @@ import Mixtape from "../types/mixtape";
 import {
     Card, CardActionArea, CardActions, CardContent,
     Container,
-    IconButton,
-    ListItemIcon,
-    Menu,
-    MenuItem,
     Typography
 } from "@mui/material";
-import {Close as CloseIcon, Edit as EditIcon, MoreVert as MoreVertIcon} from "@mui/icons-material";
-import React, {useEffect} from "react";
 import MixtapeUtils from "../utils/mixtape-utils";
-import MixtapeForm from "./MixtapeForm";
-import ConfirmDialog from "./ConfirmDialog";
-import useConfirmDialog from "../hooks/useConfirmDialog";
-import {MixtapeApi} from "../api/mixify-api";
-import {toast} from "react-toastify";
-import useForm from "../hooks/useForm";
-import useMenu from "../hooks/useMenu";
 import {Link, useNavigate} from "react-router-dom";
 import CardImageWithPlayButton from "./CardImageWithPlayButton";
-import {useBackdrop} from "../context/backdropContext";
+import MixtapeMenu from "./MixtapeMenu";
 
 export default function MixtapeCard({mixtape, onEdit, onDelete}: {
     mixtape: Mixtape,
@@ -28,30 +15,6 @@ export default function MixtapeCard({mixtape, onEdit, onDelete}: {
     onDelete: (deletedMixtape: Mixtape) => void,
 }) {
     const navigate = useNavigate();
-    const {menuAnchorEl: mixtapeAnchorEl, isMenuOpen: isMixtapeMenuOpen, openMenu: openMixtapeMenu, closeMenu: closeMixtapeMenu} = useMenu();
-    const {isFormOpen: isMixtapeFormOpen, openForm: openMixtapeForm, closeForm: closeMixtapeForm} = useForm();
-    const {enableBackdrop} = useBackdrop();
-
-    const {
-        isConfirmDialogOpen: isDeleteConfirmDialogOpen,
-        openConfirmDialog: openDeleteConfirmDialog,
-        closeConfirmDialog: closeDeleteConfirmDialog
-    } = useConfirmDialog();
-
-    const mixtapeMenuId = `mixtape-${mixtape.id}-menu`;
-
-    useEffect(() => {
-        enableBackdrop(isMixtapeMenuOpen);
-    }, [isMixtapeMenuOpen, enableBackdrop])
-
-    const handleDeleteConfirmed = async () => {
-        await MixtapeApi.deleteMixtape(mixtape);
-
-        onDelete(mixtape);
-        toast.success("Successfully deleted mixtape.");
-
-        closeDeleteConfirmDialog();
-    };
 
     return (
         <Card elevation={5} sx={{display: "flex", position: "relative"}}>
@@ -82,58 +45,8 @@ export default function MixtapeCard({mixtape, onEdit, onDelete}: {
             </CardActionArea>
 
             <CardActions>
-                <IconButton onClick={(e) => openMixtapeMenu(e.currentTarget)}
-                            aria-controls={isMixtapeMenuOpen ? mixtapeMenuId : undefined}
-                            aria-haspopup="true"
-                            aria-expanded={isMixtapeMenuOpen ? 'true' : undefined}
-                            sx={{
-                                alignSelf: "flex-start",
-                                position: "absolute",
-                                top: (theme) => theme.spacing(1),
-                                right: 0,
-                            }}>
-                    <MoreVertIcon/>
-                </IconButton>
-                <Menu
-                    id={mixtapeMenuId}
-                    anchorEl={mixtapeAnchorEl}
-                    open={isMixtapeMenuOpen}
-                    onClose={closeMixtapeMenu}
-                    onClick={closeMixtapeMenu}
-                >
-                    <MenuItem onClick={openMixtapeForm}>
-                        <ListItemIcon>
-                            <EditIcon fontSize="small"/>
-                        </ListItemIcon>
-                        Edit
-                    </MenuItem>
-                    <MenuItem onClick={openDeleteConfirmDialog}>
-                        <ListItemIcon>
-                            <CloseIcon fontSize="small"/>
-                        </ListItemIcon>
-                        Delete
-                    </MenuItem>
-                </Menu>
+                <MixtapeMenu mixtape={mixtape} onEdit={onEdit} onDelete={onDelete} />
             </CardActions>
-
-            {isDeleteConfirmDialogOpen &&
-              <ConfirmDialog
-                open={isDeleteConfirmDialogOpen}
-                title={`Do you really want to delete your "${mixtape.title}" mixtape?`}
-                onCancel={closeDeleteConfirmDialog}
-                onConfirm={handleDeleteConfirmed}
-              />
-            }
-
-            {isMixtapeFormOpen &&
-              <MixtapeForm
-                title="Edit mixtape"
-                open={isMixtapeFormOpen}
-                mixtape={mixtape}
-                onSave={onEdit}
-                onClose={closeMixtapeForm}
-              />
-            }
         </Card>
     );
 }
