+++ frontend/src/components/MixtapeDetails.tsx
@@ -1,23 +1,13 @@
 import Mixtape from "../types/mixtape";
-import ConfirmDialog from "./ConfirmDialog";
-import MixtapeForm from "./MixtapeForm";
 import {
     Box,
     Container,
-    IconButton, ListItemIcon, Menu, MenuItem,
     Typography
 } from "@mui/material";
-import React, {useEffect} from "react";
-import useMenu from "../hooks/useMenu";
-import useForm from "../hooks/useForm";
-import useConfirmDialog from "../hooks/useConfirmDialog";
-import {MixtapeApi} from "../api/mixify-api";
-import {toast} from "react-toastify";
 import MixtapeUtils from "../utils/mixtape-utils";
-import {Close as CloseIcon, Edit as EditIcon, MoreVert as MoreVertIcon} from "@mui/icons-material";
 import CardImageWithPlayButton from "./CardImageWithPlayButton";
-import {useBackdrop} from "../context/backdropContext";
 import {useNavigate} from "react-router-dom";
+import MixtapeMenu from "./MixtapeMenu";
 
 export default function MixtapeDetails({mixtape, onEdit, onDelete}: {
     mixtape: Mixtape,
@@ -25,30 +15,6 @@ export default function MixtapeDetails({mixtape, onEdit, onDelete}: {
     onDelete: (deletedMixtape: Mixtape) => void,
 }) {
     const navigate = useNavigate();
-    const {menuAnchorEl: mixtapeMenuAnchorEl, isMenuOpen: isMixtapeMenuOpen, openMenu: openMixtapeMenu, closeMenu: closeMixtapeMenu} = useMenu();
-    const {isFormOpen: isMixtapeFormOpen, openForm: openMixtapeForm, closeForm: closeMixtapeForm} = useForm();
-    const {enableBackdrop} = useBackdrop();
-
-    const {
-        isConfirmDialogOpen: isDeleteConfirmDialogOpen,
-        openConfirmDialog: openDeleteConfirmDialog,
-        closeConfirmDialog: closeDeleteConfirmDialog
-    } = useConfirmDialog();
-
-    const mixtapeMenuId = `mixtape-${mixtape.id}-menu`;
-
-    useEffect(() => {
-        enableBackdrop(isMixtapeMenuOpen);
-    }, [enableBackdrop, isMixtapeMenuOpen]);
-
-    const handleDeleteConfirmed = async () => {
-        await MixtapeApi.deleteMixtape(mixtape);
-
-        onDelete(mixtape);
-        toast.success("Successfully deleted mixtape.");
-
-        closeDeleteConfirmDialog();
-    };
 
     return (
         <Container sx={{display: "flex", flexDirection: "column", gap: 2, p: 0}}>
@@ -65,57 +31,7 @@ export default function MixtapeDetails({mixtape, onEdit, onDelete}: {
                     <Typography>{MixtapeUtils.formatNumberOfTracks(mixtape.tracks)}</Typography>
                 </Container>
 
-                <IconButton onClick={(e) => openMixtapeMenu(e.currentTarget)}
-                            aria-controls={isMixtapeMenuOpen ? mixtapeMenuId : undefined}
-                            aria-haspopup="true"
-                            aria-expanded={isMixtapeMenuOpen ? 'true' : undefined}
-                            sx={{
-                                alignSelf: "flex-start",
-                                position: "absolute",
-                                top: "-8px",
-                                right: "-16px",
-                            }}>
-                    <MoreVertIcon/>
-                </IconButton>
-                <Menu
-                    id={mixtapeMenuId}
-                    anchorEl={mixtapeMenuAnchorEl}
-                    open={isMixtapeMenuOpen}
-                    onClose={closeMixtapeMenu}
-                    onClick={closeMixtapeMenu}
-                >
-                    <MenuItem onClick={openMixtapeForm}>
-                        <ListItemIcon>
-                            <EditIcon fontSize="small"/>
-                        </ListItemIcon>
-                        Edit
-                    </MenuItem>
-                    <MenuItem onClick={openDeleteConfirmDialog}>
-                        <ListItemIcon>
-                            <CloseIcon fontSize="small"/>
-                        </ListItemIcon>
-                        Delete
-                    </MenuItem>
-                </Menu>
-
-                {isDeleteConfirmDialogOpen &&
-                  <ConfirmDialog
-                    open={isDeleteConfirmDialogOpen}
-                    title={`Do you really want to delete your "${mixtape.title}" mixtape?`}
-                    onCancel={closeDeleteConfirmDialog}
-                    onConfirm={handleDeleteConfirmed}
-                  />
-                }
-
-                {isMixtapeFormOpen &&
-                  <MixtapeForm
-                    title="Edit mixtape"
-                    open={isMixtapeFormOpen}
-                    mixtape={mixtape}
-                    onSave={onEdit}
-                    onClose={closeMixtapeForm}
-                  />
-                }
+                <MixtapeMenu mixtape={mixtape} onEdit={onEdit} onDelete={onDelete}/>
             </Container>
 
             <Box>
